Take two on handling module refactor.
